When a job is enqueued from another job, relate the two nodes
I'm sure there's a better way to know if the job is being enqueued from
another job process but for now, I'm leveraging the rather sloppy
solution of checking ActiveJob's log formatter.

The default formatter, when running from a job process, looks something
like:

```
["ActiveJob", "ActiveJob::EnqueuedJob", "3356850e-5b35-497d-aacb-5a559603e981"]
```

By naively taking `.last` and looking in our graph, we can successfully
create a relationship between two nodes if there's a node with that
`job_id`. This is a brittle solution that breaks down as soon as that
log formatter changes.